Starting to rework docs around the secret resource
This is the start of a reworking of our docs around the
secrets API. There will be a number of follow-up patches
before this document will be considered complete.
Until this document is considered complete, I'm not adding
this page to the table of contents. Also, to limit
individual CR size, I've tossed TODOs on the areas that
will be worked on in future CRs.

Change-Id: I8599628d51f4b5aae9ec1b8fd76f419d5b38bdcc