Removing the debian folder
On Debian systems, we require upstream to *not* want upstream
to add a debian folder on their upstream code. This goes against
all the worflows we have for maintaining packages inside a specific
branch which contains that debian folder.

As I am about to package Barbican in Debian (reminder: I'm the
package maintainer for everything else in OpenStack in Debian), I
need this folder to go away, so I can do my packaging work. I
certainly appreciate to have a template package to start with though,
so thanks to whoever wrote that (it will appear in debian/copyright).

Change-Id: I1adadb4b95345308f613f3a52b9fffdbc9229bee