Switching how we handle sessions in p11_crypto
Previously, we were trying to use pre-existing sessions and
semaphores to deal with our concurrency issues. However, this
led to HSM memory exhaustion due to the sessions never closing.
To mitigate this, I'm switching to a model where we're opening
and closing sessions as we need them. By extension, this removes
our need for semaphores. We can explore ways of making this more
efficient in the future, but this change should address the
outstanding issue.

Fixes Bug: 1436529

Change-Id: I830f3f770a93d5d82650ec93bb8fa89e3cee5a73