Add new smoke tests for RSA type containers and secrets
This adds new functional tests that mimic the curl test
examples in doc/source/api/reference/secret_types.rst.

The new tests are:
  Create private secret with POST.
  Create private secret with POST and PUT.
  Create public secret with POST.
  Create public secret with POST and PUT.
  Create passphrase secret with POST.
  Create RSA container with POST.
  Order RSA container
  Order certifciates

The secrets used in this test are generated using OpenSSL.
Per the API reference, the stored and retrieved
secret values should match.

The secret formats are:
  private: PKCS#8 PEM
  public: PEM
  passphrase: text
  certificate: PEM

Secrets sent in JSON via POST are base64 encoded
Secrets sent in the body via POST and PUT are sent natively

The public and private tests fail pending a fix for 1441866.

Change-Id: I82d9864e0ee6e20d1f07b0f171c776ac4ea24171