Fix KMIP Secret Store input/output
Fix KMIP Secret Store inputs and outpus.  Barbican core will pass in PEM
secrets as Base64(PEM), so the secret store has to convert them to DER
for use.  On secret retrieval the secret store is expected to return
Base64(PEM) back to Barbican, so we need to convert the DER coming back
from the KMIP device into a PEM.

Change-Id: I79693be16fb6ab3b781d6f215daf30274fc8200e