Fix Dogtag plugin to make latest functional tests pass
For generated keys, we add a field in the Dogtag metadata
"generated" to let the Dogtag plugin know to base64.b64decode()
the data when it is returned.  This will not be needed once the
Dogtag API is modified to automatically encode the stored results.

Also, corrected a check for existing encrypted data on a secret PUT
operation.  The existing check was only valid for the crypto_store plugin.

Change-Id: I4ca1d4f144eb740e6936c88a3786e91ee37d8f93