Add project_id to Secret model
The approach of having a separate table for the Secret-Project
associations is being deprecated in favor of having a foreign key to the
Project model (as described in the blueprint). This commit is an initial
step towards that goal, so the following was done:

* Add the project_id to the Secret model.
* Enhance the logic in the repositories and plugin.resources in order to
  first search that new field, and if it doesn't work, fall back to the
  previous approach.
* Added the necessary code for the project_id to be filled for the
  secret.

Note that the association table approach still works and will be removed
in a subsequent commit.

blueprint data-remove-tenant-secret-assoc

Change-Id: Icabeb5b3fc2e5aaa8c0c4fc4dfb5aae7405c4e2c