Add Barbican configs for SQLAlchemy pool settings
In order to support tuning of database operation in production, this
commit exposes pool-related SQLAlchemy parameters. This includes
enabling DEBUG/INFO logging from the pool (when it connects, how many
connections are active), the number of connection instances to keep,
the maximum number of connections over that to allow, and the
SQLAlchemy pool class to use.

Change-Id: I573147d16f61966eb74479b66077efa23c638a43