Fix NewRelic error reporting
Adding a hook to capture errors for NewRelic. Unfortunately,
the tracebacks reported to NewRelic do not completely reflect
the tracebacks in the logs. However, this issue has been
addressed in Pecan commit d3159f1d6bec06845b152de836793f4454d95804.
We just need to wait for a newer Pecan release to address that
issue.

Change-Id: I2cc542d09420bdf30afc3970f1bfec5a9300d72c