Changed Test Key Size to 2048
There were several places where RSA key sizes were less than 2048.
This change makes all of the keys now 2048 bits. This will help tests
with devices that are FIPS compliant and do not allow keys of 1024
bits. This is important for functional tests that use real devices.

The test code included two modules that provided example private key,
public key, and certificate. The utils module was modified to remove
its samples and the keys.py file will now contain the keys to be used
for testing. Those keys are 2048 while utils was 1024.

Change-Id: I19c5b32d7861953ebbb458a6f6336647bb13736c