Fill project_id in secrets where needed
This CR introduces a migration script that fills the project_id in the
secrets that have it set to None; Which will be the case in the secrets
created before the project_id column was introduced.

On the other hand, this also makes the project_id column in the secrets
table no longer nullable.

blueprint data-remove-tenant-secret-assoc

Change-Id: Iede2d2fcf6487ec3ef1b8d29a6000e2340d4c828