Fix gate jobs after change to infra
No longer need the envdir setting in tox.ini.

Fix unit test errors caused by new mock version

This patch corrects a few errors that is currently breaking the gate
checks.  The trigger of these errors is an upgrade to Mock, but the
root is subtle issues with how Mock was being used.  All of these
changes will work correctly with the old and new versions of Mock.

1) Fix missing attribute errors (calls to non existing assert checks)
2) Fix use of assert_has_calls() (parameter should be a list)
3) Don't assume name of Mock instance
4) Upgrade to use of newer PBR to address import errors
       (this change will be forthcoming from an infra-bot)
5) Update versions based on CR 201162
6) Add python-subunit explicitly to test-requirements.txt
7) Remove extra pip install of tempest; use version from test-requirements

Co-authored-by: Kaitlin Farr <kaitlin.farr@jhuapl.edu>
Co-authored-by: Dave McCowan <dmccowan@cisco.com>
Closes-Bug: #1475909
Change-Id: Id4f46ae77bda5a89fb112a779c089542c3987e7b