Configure dogtag installation's domain name
It seems that previously we were overriding the /etc/hosts in an
incorrect way. The gate had already several aliases for localhost which
we were completely discarding, and this might cause some issues with
some assumptions that were being taken in the gate.

On the other hand, a newer version of dogtag that is now being used in
the gate requires there to be a proper DNS domain name configured, and
this was causing the installation to fail in the gate.

So as a solution, we now force the fqdn to be localhost.localdomain. And
on the other hand, we explicitly configure dogtag to use localhost as a
hostname, since it was actually getting an alias from the hostname
command, which was not being found with the domain it was fetching from
the dnsdomainname command.

Closes-Bug: #1477118
Change-Id: I7ea518ddc35ac96fb8594d57a1a46a32f30bf3c0