Implement Models and Repositories for Resource Quotas
In the interest of smaller CRs, this CR partially implements the
quota support blueprint.  It includes code model and repository
used to store project quota information.  Unit and functional tests
are also provided to verify the implementation.

This CR also cover some small design changes, such as: the new
defaults is for unlimited quotas for all resources.

Implements: blueprint quota-support-on-barbican-resources

Change-Id: Ief79dd36fd35528c9d8586d77e01d1f9ad723f4d