Add get_count() method to project resource repositories
Getting ready to enforce project quotas, we need a way to count
the number of resources in use by a project.  This commit adds a
get_count() method for secrets, containers, orders, and consumers.

To enable the get_count() method for consumers, a new column is
added to the consumers table.  A database migration script is included
for database upgrades required by this change.

Change-Id: I7a00da218ac2980dd1b4e50c6ae58fc683ba3ee3
Partially-implements: blueprint quota-support-on-barbican-resources