Ensure a http 405 is returned on container(s) PUT
Added the template='json' option to the pecan expose decorator on
the index function on the Containers controller to ensure that we
pass control to that function rather than have pecan determine
that some content type(s) aren't what it expects.

Without this fix, pecan looks at the content type (we set as
application/json) and since its not "text/html" pecan returns an
http 406 rather than passing control on to the index function which
sets the more appropriate http 405.

Closes-Bug #1485157

Change-Id: I4ff640c7bd412b9946797d2da7be04fe78678c63