Force Identity API v3 for endpoint create in devstack
When the switch was made to the Openstack CLI, the Identity API v3 was
being taken into account. And that's the way it worked when the CR was
introduced. But for some reason, the gate is now taking v2 into account
(Which was the original behaviur).

This is problematic in the sense that commands related to identity v2
have a different syntax than commands related to identity v3.

This CR sets v3 specifically in the problematic commands, since it's the
syntax that we were considering initially. And also the goal is to move
towards the usage of Identity v3 as much as possible.

Change-Id: I5be80ce0c222be0a5439f4284351e3d30c40ef2e