Fix ca related controllers
GET /cas/ would return all the CAs even if they're not in the project.
This CR enables the behavior in which if project CAs are defined, then
it will show them when running GET /cas/; If they're not, then it will
show all CAs.

On the other hand, this fixes an issue with remove-from-project. Where
if you would remove the last CA from a project, it would still have a
preferred one. So the behavior now is:
If the user tries to delete the preferred CA; an error will be returned
unless it is the last CA left for the project.

Change-Id: I195ecc4ea6a8748e003fd17b219a727701e5e15d