Clean up CAs Policy Rules
A few policy rule names did not match between the decorators and the
policy.json file.  This commit fixes those.

Also, I changed the permissions for "get_global_preferred" to be
for service-admin only.  Project admins and users should only know
what their own preferred CA is.  Global settings are only
relavent to the overall service admin.

Also, if the first call a project admin makes is GET cas/preferred
then that call needs to create a project entry.  This commit adds the
call to do that.

Change-Id: I2e1ddbf8f2d93af5bd9b182716ad9540d7165420
Closes-bug: #1496821