Exit with error code when db_manage.py fails
The DB migration script db_manage.py always returned a 0 exit code,
even if errors were seen. This would break use of this script in
automation since the next sequence in the workflow would always be
executed even on errors. This script should return a 1 if errors are
seen. This CR adds that error code return.

Closes-bug: #1500448
Change-Id: I9f51bca409d23a90f629acb124938ecefa11208a