Remove low-level PyKMIP test asserts
This change removes two low-level asserts that test the internals of
PyKMIP core. The referenced portions of PyKMIP have been modified in the
latest release. Once the latest release has been published, these asserts
will be added back in, modified to work properly with the new version of
PyKMIP.

Partial-Bug: #1522856
Change-Id: I50b9a4ad935d6458c5dc42f8184a724c78a28a59