Remove deprecated option 'DEFAULT/verbose'
Option "verbose" from group "DEFAULT" is deprecated from the oslo.log
library. Its value may be silently ignored in the future. If this option
is not set explicitly, there is no such warning.

Furthermore, the default value of verbose is true, so there is no
need to set this value in config files.

Change-Id: If760c42da26aefe49ae52415941e6146bcbccb6a