Add missing unit test for clean_command and fix error handling
The clean_command in barbican/model/clean.py is missing
a unit test for coverage. Unit tests have been added.

There was a redundant log exception that was removed.
A check for a successfull exit was removed because there are no
apparent exceptions that raise successfull exits during clean up.

Closes-Bug: #1555182
Change-Id: Ia1e1fe07dc9081b20c5fb0bad1a979809db0be4a