Changes max string length for URL to 255
Since the barbican database character set is set to utf-8 in devstack,
the max value for a string is 255. Having more than 255 throws an error.
Fixing this bug will also allow us to use the barbican-manage command
in the devstack gate to test migration so errors like this do not
go unchecked in the future.

Change-Id: I7ea320f1640977044c8058f682b43b42e7a5e2e2
Closes-Bug: 1528774