Fix circular dependency  of certificate_manager module
The certificate_manager module had a circular dependency that made it
problematic to use for certain versions of the python interpreter.
This was because we were initializing a singleton EVENT_PLUGIN_MANAGER
that in turn loaded the plugins; The problem was that the plugins
imported certificate_manger themselves and this caused the circular
dependency. So instead of creating that class in the module, I
created a function that should be called instead.

Closes-Bug: #1570356
Change-Id: I9cdb7c3ecc9c275ee561363f6ce13248259f88d0