Implement Date Filters for Secrets
This CR implements the spec for date filters.  The only difference
between the spec and this CR is the use of the alternative ISO format
(without the "Z") for specifying the dates.  This change was made to
have a more consistent API since the Zulu designation is not used
anywhere in the API where dates are shown to the user.  Additionaly the
libraries used for date-time parsing do not make use of the Zulu
designation either.

Implements: blueprint date-filters
DocImpact
APIImpact

Change-Id: Ic8fbe3d0e8b309bb192aaddf30291d1333756064