Add seed random feature to seed HSM RNG
With the softhsm2 + openssl >= 0.9.5, barbican needs to initialise the random
seed of the random number generator (RNG).  The PKCS11 interface provides an
API call for this (C_SeedRandom) so that the caller can seed the RNG.

opensll >= 0.9.5 fails if the RNG is not seeded prior to use which causes (in
this configuration) Barbican to fail when  it needs to use the RNG in the HSM.

This change adds two configuration options in the [p11_crypto_plugin] section
called 'seed_file' and 'seed_length' which default to empty string and 32
respectively.

If the 'seed_file' configuration file is not configured then this change has no
impact.  If the seed_file is configured, then 'seed_length' bytes/chars are
read from the file and provided to eh C_SeedRandom() api call.

Change-Id: Ib063dfdac90d2a3ea9e4ec3da2d84f6ee5d45b03
Closes-Bug: #1599550