Don't inspect oslo.context
When creating an oslo.context we inspected the __init__ string to figure
out if roles were part of the context. This was added in oslo.context
2.2 and there needed to be compatibility between old and new.

Now that requirements specify oslo.context >= 2.9 we don't need this
clause any more, and it is causing release issues with older versions of
positional.

Remove the inspection and rely on newer versions of oslo.context.

Closes-Bug: #1620963
Change-Id: I2fcb4f1718ae5e5b50e26d9dee0e0df2f1e6cf72