Updates to Devstack Variables
The environment variable DEVSTACK_LOCAL_CONFIG can be set by the
gate directly or through the gate_hook.sh script.  This patch
changes gate_hook.sh to append to that variable, instead of settting
that variable, so not to overwrite any changes made by the gate
startup scripts.

Also, set the PYTHON environment variable to indicate python3
is used if indicated by the DEVSTACK_GATE variable.

Change-Id: I28f2fc0bc0a97bf52fa2ba4851a90e87ef22e992