Castellan based secret store
This will provide a Castellan based secret store, which will
allow secret stores which have a castellan backend to be used
behind barbican.  The initial example of this is the Vault
backend.

Unit tests have been added.  In local tests,
most of the functional tests do in fact pass with a local Vault
backend, though this will need to be demonstrated with a later
review which establishes a Vault based gate.

Change-Id: Ib30fb79304014592bfc37938839d60a4c10c244d