Ensure only api app initializes secret store
Right now, the multiple secret store initialization code is run
whenever the db is initialized - whether it be running the clean
db script, starting the worker, starting the keystone listener
or manage db script.  This periodically causes deadlocks when
the worker,listener and app are started soon after each other.

Its not altogether clear why the deadlock happens, but the only
table that is being written to is the secret_store table, which
has no foreign keys etc.  In any case, though, it was never the
intention that anything other than the app itself initialize the
secret stores from its config file.

This patch makes sure that happens.

Change-Id: I711b91b19b9d65260a21b41d6f9e18b9e282138a
Closes-bug: 1738863