Import octavia-v2-dsvm-py35-scenario
Import from project-config, this job was run on master only but is now
run in octavia on newer branches. Import it and let it thus run
on master and future stable branches.
Also updates to using the v2 API scenario tests as these are the go
forward tempest tests for Octavia.

Change-Id: Iea049fac666788d394d6e1abc053493d2c7fd216