tox: Stop building *all* docs in 'docs'
The 'docs' target currently builds the documentation trees in 'api-ref' and
'api-guide', in addition to 'doc'. This massively increases the amount of
time docs take to build both locally and in the gate. It's not necessary for
gate, since separate jobs take care of the other documents for barbican.
As such, we should stop doing it.

For users that *do* care about this (for whatever reason) a new 'all-docs'
target is included.

Change-Id: I8b65c16d7bf574f045b702de853ad81389f140d5