Support listener pooling in keystone listener
some oslo.messaging transport backends support listener pooling,
when the listener declares a pool it belongs, and each message of
the topic is delivered once to each pool (plus to 'default' pool for
those that have not declared a pool).
In particular, rabbitmq and kafka support it.
This mechanism should be preferrable over setting a separate topic
for barbican to listen notifications on.

This patch adds a new config option `[keystone_notifications]pool_name`
(None by default for backward compatibility) that can be used to set
up barbican-keystone-listener to create a separate pool for its
listeners.

For more details see oslo.messaging docs
https://docs.openstack.org/oslo.messaging/latest/reference/notification_listener.html

Change-Id: Ie011266f5ebe03bc6053bfe68e2bee27e07ea11c