PDF Documentation Build tox target
This patch adds a `pdf-docs` tox target that will build
PDF versions of our docs. As per the Train community goal:

  https://governance.openstack.org/tc/goals/selected/train/pdf-doc-generation.html

Add sphinxcontrib-svg2pdfconverter to doc/requirements.txt
to convert our SVGs.

Also, due to the opposing graphic format, the .gif file
had to be converted (not renamed) to .png so the LaTeX
builder would accept and read the images.

Change-Id: I6384e56222991d39f4fd1bc99e82be6c89fb16a3
Story: 2006072