Set db_auto_create default to False
Change the default value for the db_auto_create option to False. This is
vital as the flag is managing upgrades as well if the databases already
exist.

It will prevent production deployments from having their databases
impacted if an API daemon is started for any reason pointing to
a production database.

Change-Id: Id7eac78737af76afe628deeca7c15c2ac969d47e