Gate on py3
A local devstack test with

  USE_PYTHON3=True
  DEVSTACK_GATE_USE_PYTHON3=True

and

  PROJECTS="openstack/barbican $PROJECTS"
  enable_plugin barbican https://opendev.org/openstack/barbican

blew up because uwsgi tried to run under py2.

The dep sets USE_PYTHON3 to True by default, which may or may not be
sufficient to expose $problems, but let's see.

[Later] We think the (legacy) grenade-devstack-barbican job was failing
because it was trying to "upgrade" from py2 to py3 and running into
packaging conflicts. This explicitly sets DEVSTACK_GATE_USE_PYTHON3=True
in that job to resolve the problem.

Depends-On: https://review.opendev.org/649097
Change-Id: I6aabb1463b001e81e1e6f7d44fb66e9b7ff3a251