Make broken fedora_latest job n-v
barbican-dogtag-devstack-functional-fedora-latest is broken
for all stable branch till stable/stein.

That job was made n-v in train cycle along with other updates
but that was not backported to other stable branches.
- https://review.opendev.org/#/c/662543/16

I do not know if all the changes of above patch can be backported
which still does not make this job working, so i am making this job
n-v only and later we can backport the complete working fix from master.

Change-Id: I37e6c5a5dbd913ad7ef263f1c5aac9114f5d3ca9