Rebase alembic migrations
This patch squashes all database migrations up to and including
the Ocata release into a single migration.

By squashing the migrations into a single one, we are able
to fix the migration issues in MySQL 8.0+.

There has only been one database migration since Ocata, which
was not changed, and any existing database will be compatible
with this change.

This patch also unblocks testing in Ubuntu Focal, which uses
MySQL 8.0 as its default database.

Change-Id: I66c4c5dc91ac3fe486784600d4f58ef4ddb8484c
Story: 2007732
Task: 39896