Allow multiple token labels for PKCS#11 driver
This patch changes the slot selection logic to look for more than one
token label in the list of slots, using the first token that is found.

This change is required to enable load balancing with devices that
use separate tokens for this feature.

This patch adds a new option in the [p11_crypto_plugin] section
"token_labels", and deprecates the "token_label" option.

For backwards compatibility, the "token_label" option will still be used
if present.

Change-Id: Ic2b85246c37e856c38cb47613313b19e653118de