Add ACL default to allow project read
Allow a project to read secrets and containers by default.  This is
needed when those objects have no ACL in the database.  The policy enforcer
defaults to False for missing values. [1]  So, attempting to check the
read_project_access for a secret or container will always return False.

If a secret does have an ACL, we'll still return whatever is stored
in the database.

[1] https://opendev.org/openstack/oslo.policy/src/branch/master/oslo_policy/_checks.py#L335-L338

Change-Id: I584f7b67f2f95caa7c4db3d9d9222d0a9d38442d