Fix alembic migrations
A change was introduced to SQL Alchemy 1.4.8 that breaks our Alembic
migrations.  Specifically it breaks the "add_secret_consumers" migration
because it attempts to check for a table using an object that is no
longer allowed to call that function.

This patch removes the bespoke validation, because alembic should take
care of checking the schema.

Story: 2008967
Task:  42606

Change-Id: I36fb10445413fb1ec4046ab6c2525eae47d85ea1