Fix naming in secret meta tests
The create_secret() helper function returns the tuple
(secret_uuid, http_response), but the tests confusingly unpack
the values into varialbes that are flipped. i.e. the UUID was unpacked
into a secret_resp variable and the response was unpacked into a
secret_uuid variable.

This patch fixes the order of names so the variable names actually reflect
the value they contain.

Change-Id: Iece4b91d7cbc5559645bb83dd158753926e2ba4a